git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/branches/feature/csharpbinding@28063 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/proctex/ptpdlight/ptpdlight_generic.cpp
@@ -324,7 +324,7 @@ void ProctexPDLight::Animate_Generic ()
       Lumel* scratchPtr = scratch + 
         mapTile.tilePartY * (blitPitch / sizeof (Lumel)) +
         mapTile.tilePartX;
-      int scratchPitch = (blitPitch / sizeof (Lumel)) - mapW;
+      size_t scratchPitch = (blitPitch / sizeof (Lumel)) - mapW;
 
       csRGBcolor mapMax = mapTile.maxValue;
       mapMax.red   = lutR[mapMax.red]   >> shiftR;
